jobs: clean up retry error handling
Also moves the code to a new file.

Release justification: category 2

Release note: None
