+++ pkg/jobs/registry_test.go
@@ -622,7 +622,7 @@ func TestRetriesWithExponentialBackoff(t *testing.T) {
 		expectedResumed := int64(0)
 		runTest(t, jobID, retryCnt, expectedResumed, lastRun, &bti, func(_ int64) {
 			<-bti.resumeCh
-			bti.errCh <- NewRetryJobError("injecting error to retry running")
+			bti.errCh <- MarkAsRetryJobError(errors.New("injecting error to retry running"))
 			<-bti.transitionCh
 		})
 	})
@@ -669,13 +669,13 @@ func TestRetriesWithExponentialBackoff(t *testing.T) {
 		<-bti.resumeCh
 		bti.errCh <- errors.Errorf("injecting error to revert")
 		<-bti.failOrCancelCh
-		bti.errCh <- NewRetryJobError("injecting error in reverting state to retry")
+		bti.errCh <- MarkAsRetryJobError(errors.New("injecting error in reverting state to retry"))
 		<-bti.transitionCh
 		expectedResumed := bti.resumed.Count()
 		retryCnt := 1
 		runTest(t, jobID, retryCnt, expectedResumed, lastRun, &bti, func(_ int64) {
 			<-bti.failOrCancelCh
-			bti.errCh <- NewRetryJobError("injecting error in reverting state to retry")
+			bti.errCh <- MarkAsRetryJobError(errors.New("injecting error in reverting state to retry"))
 			<-bti.transitionCh
 		})
 	})
@@ -692,12 +692,12 @@ func TestRetriesWithExponentialBackoff(t *testing.T) {
 		pauseOrCancelJob(t, ctx, bti.kvDB, bti.registry, jobID, cancel)
 		bti.errCh <- nil
 		<-bti.failOrCancelCh
-		bti.errCh <- NewRetryJobError("injecting error in reverting state")
+		bti.errCh <- MarkAsRetryJobError(errors.New("injecting error in reverting state"))
 		expectedResumed := bti.resumed.Count()
 		retryCnt := 1
 		runTest(t, jobID, retryCnt, expectedResumed, lastRun, &bti, func(retryCnt int64) {
 			<-bti.failOrCancelCh
-			bti.errCh <- NewRetryJobError("injecting error in reverting state")
+			bti.errCh <- MarkAsRetryJobError(errors.New("injecting error in reverting state"))
 			waitUntilCount(t, bti.jobMetrics.FailOrCancelRetryError, retryCnt+1)
 		})
 	})
@@ -719,7 +719,7 @@ func TestRetriesWithExponentialBackoff(t *testing.T) {
 		<-bti.resumeCh
 		bti.errCh <- errors.Errorf("injecting error to revert")
 		<-bti.failOrCancelCh
-		bti.errCh <- NewRetryJobError("injecting error in reverting state to retry")
+		bti.errCh <- MarkAsRetryJobError(errors.New("injecting error in reverting state to retry"))
 		<-bti.transitionCh
 		expectedResumed := bti.resumed.Count()
 		retryCnt := 1
@@ -730,7 +730,7 @@ func TestRetriesWithExponentialBackoff(t *testing.T) {
 			// failed regardless of the fact that it is currently pause-requested in the
 			// jobs table. This is because we currently do not check the current status
 			// of a job before marking it as failed.
-			bti.errCh <- NewRetryJobError("injecting error in reverting state to retry")
+			bti.errCh <- MarkAsRetryJobError(errors.New("injecting error in reverting state to retry"))
 			<-bti.transitionCh
 			waitUntilStatus(t, bti.tdb, jobID, StatusPaused)
 			require.NoError(t, bti.registry.Unpause(ctx, nil, jobID))
